test: Add comprehensive unit tests for rx_bus manager and OneWire driver
Implements complete test coverage for the rx_bus abstraction layer:

- test_rx_bus_manager: 36 tests covering initialization, bus registration,
  removal, lookup, thread-safe callbacks, and command pattern

- test_rx_bus_onewire: 55 tests covering initialization, reset/presence
  detection, bit/byte/buffer operations, ROM commands, search algorithm,
  and GPIO error injection

Key features:
- Mock hardware infrastructure (CMT timer, system registers, GPIO, CRC)
- Auto-incrementing timer mock prevents infinite loops in delay functions
- Mock headers shadow real hardware headers for host-side testing
- All tests pass (91 total tests across 13 test suites)

Resolves #61

+++ star-rx72n-firmware/tests/mocks/mock_rx_crc.c
@@ -11,6 +11,7 @@
 #include "mock_rx_crc.h"
 
 #include <stdbool.h>
+#include <stddef.h>
 
 /* =============================================================================
  * Constants
@@ -54,7 +55,7 @@ void mock_crc8_set_override(bool enable)
  * =============================================================================
  */
 
-uint8_t rx_crc8_maxim(const uint8_t* data, size_t length)
+uint8_t rx_crc8_maxim(const uint8_t* data, uint32_t length)
 {
   if (s_override_enabled) {
     return s_mock_crc_value;
@@ -66,7 +67,7 @@ uint8_t rx_crc8_maxim(const uint8_t* data, size_t length)
 
   uint8_t crc = 0;
 
-  for (size_t i = 0; i < length; ++i) {
+  for (uint32_t i = 0; i < length; ++i) {
     crc ^= data[i];
 
     for (uint8_t j = 0; j < k_bits_per_byte; ++j) {
